[refactor] #62 dto패키지 분리
