+++ src/main/java/kyonggi/bookslyserver/domain/shop/controller/ShopApiController.java
@@ -1,9 +1,10 @@
 package kyonggi.bookslyserver.domain.shop.controller;
 
 
+
 import kyonggi.bookslyserver.domain.shop.dto.request.ShopCreateRequestDto;
-import kyonggi.bookslyserver.domain.shop.dto.response.ShopCreateResponseDto;
-import kyonggi.bookslyserver.domain.shop.dto.response.ShopRegisterDto;
+import kyonggi.bookslyserver.domain.shop.dto.response.shop.ShopCreateResponseDto;
+import kyonggi.bookslyserver.domain.shop.dto.response.shop.ShopRegisterDto;
 import kyonggi.bookslyserver.domain.shop.service.ShopService;
 import kyonggi.bookslyserver.global.common.SuccessResponse;
 import lombok.RequiredArgsConstructor;
