+++ src/main/java/kyonggi/bookslyserver/domain/shop/dto/response/menu/MenuCreateResponseDto.java
@@ -1,0 +1,9 @@
+package kyonggi.bookslyserver.domain.shop.dto.response.menu;
+
+import lombok.Builder;
+
+@Builder
+public record MenuCreateResponseDto(
+        Long id
+) {
+}
