+++ src/main/java/kyonggi/bookslyserver/domain/shop/dto/response/shop/ShopCreateResponseDto.java
@@ -1,4 +1,4 @@
-package kyonggi.bookslyserver.domain.shop.dto.response;
+package kyonggi.bookslyserver.domain.shop.dto.response.shop;
 
 import kyonggi.bookslyserver.domain.shop.dto.BusinessScheduleDto;
 import kyonggi.bookslyserver.domain.shop.entity.Shop.Shop;
